requirements.txt
streamlit
openai

pysqlite3-binary
PyPDF2
protobuf==3.20
chromadb

request
json